refactor: refactor DID controller implementation to use account-based storage
