+++ x/did/controller/accstd/account.go
@@ -1,0 +1,1 @@
+package accstd
