+++ x/did/controller/accstd/address.go
@@ -1,4 +1,4 @@
-package controller
+package accstd
 
 import (
 	"crypto/ecdsa"
