+++ x/did/controller/accstd/pubkey.go
@@ -1,4 +1,4 @@
-package controller
+package accstd
 
 import (
 	"strings"
