+++ x/did/controller/accstd/secpk251.go
@@ -1,4 +1,4 @@
-package controller
+package accstd
 
 import (
 	"bytes"
