--- x/did/controller/spawn.go
@@ -1,1 +1,0 @@
-package controller
