+++ x/did/controller/wallet/actor.go
@@ -1,0 +1,1 @@
+package wallet
