+++ x/did/controller/wallet/init.go
@@ -1,0 +1,1 @@
+package wallet
