+++ x/did/controller/wallet/signer.go
@@ -1,4 +1,4 @@
-package controller
+package wallet
 
 import (
 	"context"
