+++ x/did/controller/wallet/txns.go
@@ -1,0 +1,1 @@
+package wallet
