+++ x/did/controller/wallet/wallet.go
@@ -1,0 +1,1 @@
+package wallet
